Add routineBackups, taskBackups and taskExecutors
This adds functionality for the two new database models to be used with the DataProvider.
This also caches a TaskExecutor object for each task which is required to make the TaskView show properly after Backups with the new "Last Backup" indicator.
This is still a bit messy, but working for now.
The process of getting the "Last Backup" indicator was rather a pain.
This might get refactored in the future.