Create utils for Backup indication
This adds time_utils to perform actions with timestamps, to properly display information
This also adds last_backup_indicator which is a widget, showing the last time a Backup was performed and whether you should create another one.
In future, you should be able to set this time to whatever you want.